config.py
import os
from dotenv import load_dotenv
import logging
load_dotenv()
CHROME_PROFILE_NAME = "Profile 2"
# API Keys
GEMINI_API_KEY = os.getenv('GEMINI_API_KEY')
if not GEMINI_API_KEY:
    raise ValueError("GEMINI_API_KEY not found in .env file. Please get a key from Google AI Studio and add it to your .env file.")

# Wake Word
WAKE_WORD = "eva"
GOODBYE_PHRASE = "goodbye eva"

# Paths
BASE_DIR = os.path.dirname(os.path.abspath(__file__))
MODEL_WEIGHTS_DIR = os.path.join(BASE_DIR, 'models', 'model_weights')
C_EXECUTOR_DIR = os.path.join(BASE_DIR, 'execution', 'c_executors')
SCREENSHOT_TEMP_DIR = os.path.join(BASE_DIR, 'temp_screenshots')
LOG_DIR = os.path.join(BASE_DIR, 'logs')

# Whisper Model Settings
WHISPER_MODEL_SIZE = "large"  # ✅ Changed to large (methodology)
WHISPER_DEVICE = "cpu"
WHISPER_COMPUTE_TYPE = "int8"

LOG_LEVEL = logging.INFO

# Gemini Settings (for screen summary + coordinate filtering ONLY)
GEMINI_MODEL = "gemini-2.0-flash"
GEMINI_FALLBACK_MODELS = [
    "gemini-2.0-flash",
    "gemini-2.0-flash-exp",
    "gemini-1.5-pro",
    "gemini-1.5-flash"
]
GEMINI_TEMPERATURE = 0.3
GEMINI_MAX_RETRIES = 3

# Classification Settings
CLASSIFICATION_CONFIDENCE_THRESHOLD = 0.6

# Audio Settings
CHUNK_SIZE = 1024
SAMPLE_RATE = 16000
RECORD_SECONDS = 5

# Session Settings
SESSION_TIMEOUT = 10  # seconds
SMTP_HOST = "smtp.gmail.com"
SMTP_PORT = 587
# SMTP credentials provided by user. WARNING: storing plaintext passwords in source is insecure.
# Prefer setting these via environment variables or a .env file and using load_dotenv().
SMTP_USER = "sabnisshriya7@gmail.com"
SMTP_PASSWORD = "esyk zboz rdmw qfjr"
SMTP_USE_TLS = True
